requirements.txt
Django>=5.0
psycopg2>=2.9
langchain
langchain-ollama
langchain_community 
langchainhub
djangorestframework
django-encrypted-json-fields==1.0.4
django-json-widget
transformers
tiktoken
redis
celery
qdrant-client